requirements.txt

viam-sdk==0.46.0

typing-extensions
